+++ client/testsources/v1/testsources.go
@@ -78,21 +78,24 @@ func (s TestSourcesClient) Get(name string) (*testsourcev1.TestSource, error) {
 
 // Create creates new test source CRD
 func (s TestSourcesClient) Create(testSource *testsourcev1.TestSource, options ...Option) (*testsourcev1.TestSource, error) {
-	secrets := make(map[string]string, 0)
-	for _, option := range options {
-		for key, value := range option.Secrets {
-			secrets[key] = value
+	if len(options) != 0 {
+		secrets := make(map[string]string, 0)
+		for _, option := range options {
+			for key, value := range option.Secrets {
+				secrets[key] = value
+			}
+		}
+
+		secretName := secret.GetMetadataName(testSource.Name, secretKind)
+		if len(secrets) != 0 {
+			if err := s.secretClient.Create(secretName, testSource.Labels, secrets); err != nil {
+				return nil, err
+			}
+
+			updateTestSourceSecrets(testSource, secretName, secrets)
 		}
 	}
 
-	secretName := secret.GetMetadataName(testSource.Name, secretKind)
-	if len(secrets) != 0 {
-		if err := s.secretClient.Create(secretName, testSource.Labels, secrets); err != nil {
-			return nil, err
-		}
-	}
-
-	updateTestSourceSecrets(testSource, secretName, secrets)
 	if err := s.k8sClient.Create(context.Background(), testSource); err != nil {
 		return nil, err
 	}
@@ -102,25 +105,30 @@ func (s TestSourcesClient) Create(testSource *testsourcev1.TestSource, options .
 
 // Update updates test source
 func (s TestSourcesClient) Update(testSource *testsourcev1.TestSource, options ...Option) (*testsourcev1.TestSource, error) {
-	secrets := make(map[string]string, 0)
-	for _, option := range options {
-		for key, value := range option.Secrets {
-			secrets[key] = value
+	if len(options) != 0 {
+		secrets := make(map[string]string, 0)
+		for _, option := range options {
+			for key, value := range option.Secrets {
+				secrets[key] = value
+			}
+		}
+
+		secretName := secret.GetMetadataName(testSource.Name, secretKind)
+		if len(secrets) != 0 {
+			if err := s.secretClient.Apply(secretName, testSource.Labels, secrets); err != nil {
+				return nil, err
+			}
+
+			updateTestSourceSecrets(testSource, secretName, secrets)
+		} else {
+			if err := s.secretClient.Delete(secretName); err != nil && !errors.IsNotFound(err) {
+				return nil, err
+			}
+
+			clearTestSourceSecrets(testSource, secretName)
 		}
 	}
 
-	secretName := secret.GetMetadataName(testSource.Name, secretKind)
-	if len(secrets) != 0 {
-		if err := s.secretClient.Apply(secretName, testSource.Labels, secrets); err != nil {
-			return nil, err
-		}
-	} else {
-		if err := s.secretClient.Delete(secretName); err != nil && !errors.IsNotFound(err) {
-			return nil, err
-		}
-	}
-
-	updateTestSourceSecrets(testSource, secretName, secrets)
 	if err := s.k8sClient.Update(context.Background(), testSource); err != nil {
 		return nil, err
 	}
@@ -182,6 +190,20 @@ func updateTestSourceSecrets(testSource *testsourcev1.TestSource, secretName str
 				Key:  gitTokenSecretName,
 			}
 		}
+	}
+
+	return
+}
+
+func clearTestSourceSecrets(testSource *testsourcev1.TestSource, secretName string) {
+	if testSource.Spec.Repository != nil && testSource.Spec.Repository.UsernameSecret != nil &&
+		testSource.Spec.Repository.UsernameSecret.Name == secretName {
+		testSource.Spec.Repository.UsernameSecret = nil
+	}
+
+	if testSource.Spec.Repository != nil && testSource.Spec.Repository.TokenSecret != nil &&
+		testSource.Spec.Repository.TokenSecret.Name == secretName {
+		testSource.Spec.Repository.TokenSecret = nil
 	}
 
 	return
